CLEANUP: Remove old smget
