+++ src/main/java/net/spy/memcached/ArcusClientIF.java
@@ -1328,26 +1328,6 @@ public interface ArcusClientIF extends MemcachedClientIF {
    * @param from        bkey index from
    * @param to          bkey index to
    * @param eFlagFilter element flag filter
-   * @param offset      0-base offset
-   * @param count       number of returning values. must be larger than 0.
-   *                    {@code offset + count} must not be more than 1000.
-   * @return a future that will hold the return value list of the fetch.
-   * @deprecated use {@link #asyncBopSortMergeGet(List, long, long,
-   * ElementFlagFilter, int, SMGetMode)}
-   */
-  @Deprecated
-  SMGetFuture<List<SMGetElement<Object>>> asyncBopSortMergeGet(
-          List<String> keyList, long from, long to, ElementFlagFilter eFlagFilter,
-          int offset, int count);
-
-  /**
-   * Get elements that matched both filter and bkey range criteria from
-   * multiple b+tree. The result is sorted by order of bkey.
-   *
-   * @param keyList     b+ tree key list
-   * @param from        bkey index from
-   * @param to          bkey index to
-   * @param eFlagFilter element flag filter
    * @param count       number of returning values. must be larger than 0 and not more than 1000.
    * @param smgetMode   smgetMode
    * @return a future that will hold the return value list of the fetch.
@@ -1872,26 +1852,6 @@ public interface ArcusClientIF extends MemcachedClientIF {
           String key, byte[] from, byte[] to, ElementFlagFilter eFlagFilter, int offset,
           int count, boolean withDelete, boolean dropIfEmpty,
           Transcoder<T> tc);
-
-  /**
-   * Get elements that matched both filter and bkey range criteria from
-   * multiple b+tree. The result is sorted by order of bkey.
-   *
-   * @param keyList     b+ tree key list
-   * @param from        bkey index from
-   * @param to          bkey index to
-   * @param eFlagFilter element flag filter
-   * @param offset      0-base offset
-   * @param count       number of returning values. must be larger than 0.
-   *                    {@code offset + count} must not be more than 1000.
-   * @return a future that will hold the return value list of the fetch.
-   * @deprecated use {@link #asyncBopSortMergeGet(java.util.List, byte[], byte[],
-   * ElementFlagFilter, int, SMGetMode)}
-   */
-  @Deprecated
-  SMGetFuture<List<SMGetElement<Object>>> asyncBopSortMergeGet(
-          List<String> keyList, byte[] from, byte[] to, ElementFlagFilter eFlagFilter,
-          int offset, int count);
 
   /**
    * Get elements that matched both filter and bkey range criteria from
