+++ src/main/java/net/spy/memcached/ArcusClientPool.java
@@ -1086,15 +1086,6 @@ public class ArcusClientPool implements MemcachedClientIF, ArcusClientIF {
     return this.getClient().flush(prefix, delay);
   }
 
-  @Deprecated
-  @Override
-  public SMGetFuture<List<SMGetElement<Object>>> asyncBopSortMergeGet(
-          List<String> keyList, long from, long to,
-          ElementFlagFilter eFlagFilter, int offset, int count) {
-    return this.getClient().asyncBopSortMergeGet(keyList, from, to,
-            eFlagFilter, offset, count);
-  }
-
   @Override
   public SMGetFuture<List<SMGetElement<Object>>> asyncBopSortMergeGet(
           List<String> keyList, long from, long to,
@@ -1344,15 +1335,6 @@ public class ArcusClientPool implements MemcachedClientIF, ArcusClientIF {
           Transcoder<T> tc) {
     return this.getClient().asyncBopGet(key, from, to, eFlagFilter, offset,
             count, withDelete, dropIfEmpty, tc);
-  }
-
-  @Deprecated
-  @Override
-  public SMGetFuture<List<SMGetElement<Object>>> asyncBopSortMergeGet(
-          List<String> keyList, byte[] from, byte[] to,
-          ElementFlagFilter eFlagFilter, int offset, int count) {
-    return this.getClient().asyncBopSortMergeGet(keyList, from, to,
-            eFlagFilter, offset, count);
   }
 
   @Override
