--- src/main/java/net/spy/memcached/ops/BTreeSortMergeGetOperationOld.java
@@ -1,25 +1,0 @@
-/*
- * arcus-java-client : Arcus Java client
- * Copyright 2010-2014 NAVER Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package net.spy.memcached.ops;
-
-public interface BTreeSortMergeGetOperationOld extends KeyedOperation {
-  interface Callback extends OperationCallback {
-    void gotData(String key, int flags, Object bkey, byte[] eflag, byte[] data);
-
-    void gotMissedKey(byte[] data);
-  }
-}
