+++ src/test/manual/net/spy/memcached/btreesmget/ByteArrayBKeySMGetIrregularEflagTest.java
@@ -47,38 +47,6 @@ class ByteArrayBKeySMGetIrregularEflagTest extends BaseIntegrationTest {
     testKeyList.add(key1);
     testKeyList.add(key2);
 
-    /* old SMGetIrregularEflagTest */
-    try {
-      mc.delete(key1).get();
-      mc.delete(key2).get();
-
-      mc.asyncBopInsert(key1, new byte[]{0}, eFlag, value + "0",
-              new CollectionAttributes()).get();
-      mc.asyncBopInsert(key1, new byte[]{3}, eFlag, value + "1",
-              new CollectionAttributes()).get();
-      mc.asyncBopInsert(key1, new byte[]{2}, eFlag, value + "2",
-              new CollectionAttributes()).get();
-
-      mc.asyncBopInsert(key2, new byte[]{1}, eFlag, value + "0",
-              new CollectionAttributes()).get();
-      mc.asyncBopInsert(key2, new byte[]{5}, null, value + "1",
-              new CollectionAttributes()).get();
-      mc.asyncBopInsert(key2, new byte[]{4}, eFlag, value + "2",
-              new CollectionAttributes()).get();
-
-      @SuppressWarnings("deprecation")
-      List<SMGetElement<Object>> list = mc.asyncBopSortMergeGet(
-              testKeyList, new byte[]{0}, new byte[]{10},
-              ElementFlagFilter.DO_NOT_FILTER, 0, 10).get();
-
-      for (int i = 0; i < list.size(); i++) {
-        System.out.println(list.get(i));
-      }
-    } catch (Exception e) {
-      e.printStackTrace();
-      fail(e.getMessage());
-    }
-
     try {
       mc.delete(key1).get();
       mc.delete(key2).get();
