+++ src/test/manual/net/spy/memcached/btreesmget/SMGetWithCombinationEflagTest.java
@@ -39,7 +39,6 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-@SuppressWarnings("deprecation")
 class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
 
   private final String KEY = this.getClass().getSimpleName();
@@ -76,20 +75,6 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
       fail(e.getMessage());
     }
 
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 1, 2,
-                    ElementFlagFilter.DO_NOT_FILTER, 0, 10);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.SECONDS);
-      assertTrue(map.isEmpty());
-      assertEquals(oldFuture.getMissedKeyList().size(), 10);
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
-      fail(e.getMessage());
-    }
-
     SMGetMode smgetMode = SMGetMode.UNIQUE;
     SMGetFuture<List<SMGetElement<Object>>> future = mc
             .asyncBopSortMergeGet(keyList, 1, 2,
@@ -123,26 +108,6 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
         }
       }
     } catch (Exception e) {
-      fail(e.getMessage());
-    }
-
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 0, 10,
-                    ElementFlagFilter.DO_NOT_FILTER, 0, 10);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.SECONDS);
-      assertEquals(10, map.size());
-      assertTrue(oldFuture.getMissedKeyList().isEmpty());
-
-      for (int i = 0; i < map.size(); i++) {
-        assertEquals(KEY + i, map.get(i).getKey());
-        assertEquals(i, map.get(i).getBkey());
-        assertEquals("VALUE" + i, map.get(i).getValue());
-      }
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
       fail(e.getMessage());
     }
 
@@ -188,26 +153,6 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
       fail(e.getMessage());
     }
 
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 0, 10,
-                    ElementFlagFilter.DO_NOT_FILTER, 1, 10);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.SECONDS);
-      assertEquals(10, map.size());
-      assertTrue(oldFuture.getMissedKeyList().isEmpty());
-
-      for (int i = 0; i < map.size(); i++) {
-        assertEquals(KEY + (i + 1), map.get(i).getKey());
-        assertEquals(i + 1, map.get(i).getBkey());
-        assertEquals("VALUE" + (i + 1), map.get(i).getValue());
-      }
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
-      fail(e.getMessage());
-    }
-
     SMGetMode smgetMode = SMGetMode.UNIQUE;
     SMGetFuture<List<SMGetElement<Object>>> future = mc
             .asyncBopSortMergeGet(keyList, 0, 10,
@@ -247,26 +192,6 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
         }
       }
     } catch (Exception e) {
-      fail(e.getMessage());
-    }
-
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 0, 10,
-                    ElementFlagFilter.DO_NOT_FILTER, 1, 10);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.SECONDS);
-      assertEquals(9, map.size());
-      assertTrue(oldFuture.getMissedKeyList().isEmpty());
-
-      for (int i = 0; i < map.size(); i++) {
-        assertEquals(KEY + (i + 1), map.get(i).getKey());
-        assertEquals(i + 1, map.get(i).getBkey());
-        assertEquals("VALUE" + (i + 1), map.get(i).getValue());
-      }
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
       fail(e.getMessage());
     }
 
@@ -312,26 +237,6 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
       fail(e.getMessage());
     }
 
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 0, 10,
-                    ElementFlagFilter.DO_NOT_FILTER, 1, 10);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.SECONDS);
-      assertEquals(8, map.size());
-      assertTrue(oldFuture.getMissedKeyList().isEmpty());
-
-      for (int i = 0; i < map.size(); i++) {
-        assertEquals(KEY + (i + 1), map.get(i).getKey());
-        assertEquals(i + 1, map.get(i).getBkey());
-        assertEquals("VALUE" + (i + 1), map.get(i).getValue());
-      }
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
-      fail(e.getMessage());
-    }
-
     SMGetMode smgetMode = SMGetMode.UNIQUE;
     SMGetFuture<List<SMGetElement<Object>>> future = mc
             .asyncBopSortMergeGet(keyList, 0, 10,
@@ -374,20 +279,6 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
       fail(e.getMessage());
     }
 
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 10, 0,
-                    ElementFlagFilter.DO_NOT_FILTER, 0, 10);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.SECONDS);
-      assertEquals(10, map.size());
-      assertTrue(oldFuture.getMissedKeyList().isEmpty());
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
-      fail(e.getMessage());
-    }
-
     SMGetMode smgetMode = SMGetMode.UNIQUE;
     SMGetFuture<List<SMGetElement<Object>>> future = mc
             .asyncBopSortMergeGet(keyList, 10, 0,
@@ -425,20 +316,6 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
       fail(e.getMessage());
     }
 
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 0, 10,
-                    ElementFlagFilter.DO_NOT_FILTER, 0, 10);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.SECONDS);
-      assertEquals(5, map.size());
-      assertEquals(oldFuture.getMissedKeyList().size(), 5);
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
-      fail(e.getMessage());
-    }
-
     SMGetMode smgetMode = SMGetMode.UNIQUE;
     SMGetFuture<List<SMGetElement<Object>>> future = mc
             .asyncBopSortMergeGet(keyList, 0, 10,
@@ -473,20 +350,6 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
         }
       }
     } catch (Exception e) {
-      fail(e.getMessage());
-    }
-
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 10, 0,
-                    ElementFlagFilter.DO_NOT_FILTER, 0, 10);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.SECONDS);
-      assertEquals(5, map.size());
-      assertEquals(oldFuture.getMissedKeyList().size(), 5);
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
       fail(e.getMessage());
     }
 
@@ -532,31 +395,12 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
       fail(e.getMessage());
     }
 
-    long start = System.currentTimeMillis();
-
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 0, 1000,
-                    ElementFlagFilter.DO_NOT_FILTER, 0, 500);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.MILLISECONDS);
-      // System.out.println((System.currentTimeMillis() - start) + "ms");
-    } catch (TimeoutException e) {
-      oldFuture.cancel(true);
-      fail(e.getMessage());
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
-      fail(e.getMessage());
-    }
-
     SMGetMode smgetMode = SMGetMode.UNIQUE;
     SMGetFuture<List<SMGetElement<Object>>> future = mc
             .asyncBopSortMergeGet(keyList, 0, 1000,
                     ElementFlagFilter.DO_NOT_FILTER, 500, smgetMode);
     try {
       List<SMGetElement<Object>> map = future.get(1000L, TimeUnit.MILLISECONDS);
-      // System.out.println((System.currentTimeMillis() - start) + "ms");
     } catch (TimeoutException e) {
       future.cancel(true);
       fail(e.getMessage());
@@ -588,31 +432,12 @@ class SMGetWithCombinationEflagTest extends BaseIntegrationTest {
       fail(e.getMessage());
     }
 
-    long start = System.currentTimeMillis();
-
-    /* old  SMGetWithCombinationEflagTest */
-    SMGetFuture<List<SMGetElement<Object>>> oldFuture = mc
-            .asyncBopSortMergeGet(keyList, 0, testSize,
-                    ElementFlagFilter.DO_NOT_FILTER, 0, 500);
-    try {
-      List<SMGetElement<Object>> map = oldFuture.get(1000L, TimeUnit.SECONDS);
-      // System.out.println(System.currentTimeMillis() - start + "ms");
-      assertEquals(50, map.size());
-      List<String> missed = oldFuture.getMissedKeyList();
-      assertEquals(testSize / 2, missed.size());
-    } catch (Exception e) {
-      oldFuture.cancel(true);
-      e.printStackTrace();
-      fail(e.getMessage());
-    }
-
     SMGetMode smgetMode = SMGetMode.UNIQUE;
     SMGetFuture<List<SMGetElement<Object>>> future = mc
             .asyncBopSortMergeGet(keyList, 0, testSize,
                     ElementFlagFilter.DO_NOT_FILTER, 500, smgetMode);
     try {
       List<SMGetElement<Object>> map = future.get(1000L, TimeUnit.SECONDS);
-      // System.out.println(System.currentTimeMillis() - start + "ms");
       assertEquals(50, map.size());
       Map<String, CollectionOperationStatus> missed = future.getMissedKeys();
       assertEquals(testSize / 2, missed.size());
